Minor package update
