+++ library/src/main/java/com/devbrackets/android/exomedia/core/api/MediaPlayerApi.java
@@ -14,40 +14,28 @@
  * limitations under the License.
  */
 
-package com.devbrackets.android.exomedia.core.type;
+package com.devbrackets.android.exomedia.core.api;
 
+import android.content.Context;
 import android.net.Uri;
 import android.support.annotation.FloatRange;
 import android.support.annotation.IntRange;
-import android.support.annotation.Nullable;
-import android.view.View;
 
 import com.devbrackets.android.exomedia.core.EMListenerMux;
 import com.devbrackets.android.exomedia.core.builder.RenderBuilder;
 
 /**
- * The basic APIs expected in the backing video view
+ * The basic APIs expected in the backing media player
  * implementations to allow us to create an abstraction
- * between the Native (Android) VideoView and the VideoView
+ * between the Native (Android) MediaPlayer and the AudioPlayer
  * using the ExoPlayer.
  */
-public interface VideoViewApi {
+public interface MediaPlayerApi {
+    void setDataSource(Context context, Uri uri);
 
-    interface OnSurfaceSizeChanged {
-        void onSurfaceSizeChanged(int width, int height);
-    }
+    void setDataSource(Context context, Uri uri, RenderBuilder renderBuilder);
 
-    int getHeight();
-
-    int getWidth();
-
-    void setVideoUri(@Nullable Uri uri);
-
-    void setVideoUri(@Nullable Uri uri, @Nullable RenderBuilder renderBuilder);
-
-    boolean setVolume(@FloatRange(from = 0.0, to = 1.0) float volume);
-
-    void seekTo(@IntRange(from = 0) int milliseconds);
+    void prepareAsync();
 
     boolean isPlaying();
 
@@ -57,9 +45,9 @@ public interface VideoViewApi {
 
     void stopPlayback();
 
-    void suspend();
-
     void release();
+
+    void reset();
 
     @IntRange(from = 0)
     int getDuration();
@@ -70,11 +58,15 @@ public interface VideoViewApi {
     @IntRange(from = 0, to = 100)
     int getBufferedPercent();
 
-    void setOnTouchListener(View.OnTouchListener listener);
+    int getAudioSessionId();
+
+    void setAudioStreamType(int streamType);
+
+    void setVolume(@FloatRange(from = 0.0, to = 1.0) float left, @FloatRange(from = 0.0, to = 1.0) float right);
+
+    void seekTo(@IntRange(from = 0) int milliseconds);
+
+    void setWakeMode(Context context, int mode);
 
     void setListenerMux(EMListenerMux listenerMux);
-
-    void updateAspectRatio(float aspectRatio);
-
-    void setOnSizeChangedListener(@Nullable OnSurfaceSizeChanged listener);
 }
